travel_ai_chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
motor


